migration fix

+++ Database/Migrations/2014_11_02_092851_create_permission_role_table.php
@@ -10,14 +10,17 @@ class CreatePermissionRoleTable extends Migration
      */
     public function up()
     {
-        Schema::create('permission_role', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('permission_id')->unsigned()->index();
-            $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
-            $table->integer('role_id')->unsigned()->index();
-            $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('permission_role')) 
+        {
+            Schema::create('permission_role', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('permission_id')->unsigned()->index();
+                $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
+                $table->integer('role_id')->unsigned()->index();
+                $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
